Add files via upload
I have listed a few GEN AI use cases where we are currently performing activities manually or dealing with painful tasks in our day-to-day activities. Please go through the attached GEN AI use cases. If you find any of them suitable for taking to the next level, we can work on it. Alternatively, if you think the problem needs some tuning to be eligible for the next level, we can work on that as well.